requirements.txt
asgiref==3.5.2
certifi==2026.1.4
chardet==3.0.4
click==8.1.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==4.1.13
fonttools==4.61.1
deep-translator
# httpx, httpcore will be installed by transformers/deep-translator with compatible versions
huggingface_hub>=0.20.0
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
kiwisolver==1.4.9
matplotlib==3.10.8
nltk==3.7
numpy==2.4.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
regex==2022.10.31
rfc3986==1.5.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.64.1
tzdata==2025.3
transformers
torch
pillow
pypdf>=3.0
python-docx
